+++ clickhouse_mysql/tablemigrator.py
@@ -3,7 +3,7 @@
 
 import logging
 
-from MySQLdb.cursors import SSDictCursor
+from MySQLdb.cursors import SSDictCursor,Cursor
 from clickhouse_mysql.tableprocessor import TableProcessor
 from clickhouse_mysql.tablesqlbuilder import TableSQLBuilder
 from clickhouse_mysql.event.event import Event
@@ -37,9 +37,13 @@ class TableMigrator(TableSQLBuilder):
             user=None,
             password=None,
             dbs=None,
+            schema=None,
+            distribute=None,
+            cluster=None,
             tables=None,
             tables_prefixes=None,
             tables_where_clauses=None,
+            column_skip=[],
     ):
         super().__init__(
             host=host,
@@ -47,8 +51,12 @@ class TableMigrator(TableSQLBuilder):
             user=user,
             password=password,
             dbs=dbs,
+            schema=schema,
+            distribute=distribute,
+            cluster=cluster,
             tables=tables,
             tables_prefixes=tables_prefixes,
+            column_skip=column_skip
         )
         self.client.cursorclass = SSDictCursor
 
@@ -167,17 +175,20 @@ class TableMigrator(TableSQLBuilder):
         :return: number of migrated rows
         """
 
-        self.client.cursorclass = SSDictCursor
-        self.client.connect(db=db)
 
         # build SQL statement
-        sql = "SELECT * FROM {0}".format(self.create_full_table_name(db=db, table=table))
+        full_table_name = self.create_full_table_name(db=db, table=table)
+        sql = "SELECT {0} FROM {1}".format(
+            ",".join(self.get_columns(db,full_table_name))
+            ,full_table_name)
         # in case we have WHERE clause for this db.table - add it to SQL
         if db in self.where_clauses and table in self.where_clauses[db]:
             sql += " WHERE {}".format(self.where_clauses[db][table])
 
         try:
             logging.info("migrate_table. sql={}".format(sql))
+            self.client.cursorclass = SSDictCursor
+            self.client.connect(db=db)
             self.client.cursor.execute(sql)
             cnt = 0;
             while True:
@@ -195,18 +206,26 @@ class TableMigrator(TableSQLBuilder):
                 self.chwriter.flush()
 
                 cnt += len(rows)
-        except:
-            raise Exception("Can not migrate table on host={} user={} password={} db={} table={} cnt={}".format(
-                self.host,
-                self.user,
-                self.password,
+        except Exception as ex:
+            logging.critical("Critical error: {}".format(str(ex)))
+            raise Exception("Can not migrate table on db={} table={}".format(
                 db,
                 table,
-                cnt
             ))
 
         return cnt
 
+    def get_columns(self,db,full_table_name):
+        self.client.cursorclass = Cursor
+        self.client.connect(db=db)
+        self.client.cursor.execute("DESC {}".format(full_table_name))
+        fields = []
+        for (_field, _type, _null, _key, _default, _extra,) in self.client.cursor:
+            if self.column_skip.__contains__(_field):
+                logging.debug("skip column %s",_field)
+                continue
+            fields.append(_field)
+        return fields
 
 if __name__ == '__main__':
     tb = TableBuilder(